Film controller and entities for it

+++ sakila-app/src/main/java/eu/sii/pl/sakilaapp/repository/CategoryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CategoryRepository extends JpaRepository<Category, Short> {
+public interface CategoryRepository extends JpaRepository<Category, Byte> {
 
   @Override
   List<Category> findAll();
